% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distros.R
\name{dbeta2}
\alias{dbeta2}
\title{Beta density function parameterized by mean and precision parameter}
\usage{
dbeta2(x, mu, phi)
}
\arguments{
\item{x}{Random variable}

\item{mu}{Mean of the distribnution}

\item{tau}{Precision parameter of the distribution}
}
\value{
The probability density of \code{x}
}
\description{
Beta density function parameterized by mean and precision parameter
}
\examples{
curve(dbeta2(x, 0.5, 5), 0, 1)
}
