% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distros.R
\name{dlnorm2}
\alias{dlnorm2}
\title{Log-normal density function parameterized by mean and coefficient of variation}
\usage{
dlnorm2(x, mu, tau)
}
\arguments{
\item{x}{Random variable}

\item{mu}{Mean of the distribnution}

\item{tau}{coefficient of variation of the distribution}
}
\value{
The probability density of \code{x}
}
\description{
Log-normal density function parameterized by mean and coefficient of variation
}
\examples{
curve(dlnorm2(x, 0.5,1), 0, 2)
}
