% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distros.R
\name{rbeta2}
\alias{rbeta2}
\title{Random number generator for beta distribution parameterized by mean and precision parameter}
\usage{
rbeta2(n, mu, phi)
}
\arguments{
\item{n}{Number of random values to be drawn from the distribution}

\item{mu}{Mean of the distribution}

\item{tau}{Precision parameter of the distribution}
}
\value{
A vector with \code{n} random values from the distribution
}
\description{
Random number generator for beta distribution parameterized by mean and precision parameter
}
\examples{
rbeta2(10, 0.5, 5)
}
