% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distros.R
\name{rlnorm2}
\alias{rlnorm2}
\title{Random number generator for log-normal distribution parameterized by mean and coefficient of variation}
\usage{
rlnorm2(n, mu, tau)
}
\arguments{
\item{n}{Number of random values to be drawn from the distribution}

\item{mu}{Mean of the distribution}

\item{tau}{coefficient of variation of the distribution}
}
\value{
A vector with \code{n} random values from the distribution
}
\description{
Random number generator for log-normal distribution parameterized by mean and coefficient of variation
}
\examples{
rlnorm2(10, 0.5, 0.1)
}
